# Make duk_hstring charlen lazy