--- a/src-input/duk_heap_stringtable.c
+++ b/src-input/duk_heap_stringtable.c
@@ -153,7 +153,6 @@ DUK_LOCAL duk_hstring *duk__strtable_alloc_hstring(duk_heap *heap,
 #if !defined(DUK_USE_HSTRING_ARRIDX)
 	duk_uarridx_t dummy;
 #endif
-	duk_uint32_t clen;
 
 	DUK_ASSERT(heap != NULL);
 	DUK_UNREF(extdata);
@@ -225,7 +224,6 @@ DUK_LOCAL duk_hstring *duk__strtable_alloc_hstring(duk_heap *heap,
 		DUK_HSTRING_SET_ARRIDX(res);
 		DUK_HSTRING_SET_ASCII(res);
 		DUK_ASSERT(duk_unicode_unvalidated_utf8_length(data, (duk_size_t) blen) == blen);
-		clen = blen;
 	} else {
 		/* Because 'data' is NUL-terminated, we don't need a
 		 * blen > 0 check here.  For NUL (0x00) the symbol
@@ -241,27 +239,18 @@ DUK_LOCAL duk_hstring *duk__strtable_alloc_hstring(duk_heap *heap,
 			}
 		}
 
-		clen = (duk_uint32_t) duk_unicode_unvalidated_utf8_length(data, (duk_size_t) blen);
-		DUK_ASSERT(clen <= blen);
-
 		/* Using an explicit 'ASCII' flag has larger footprint (one call site
 		 * only) but is quite useful for the case when there's no explicit
 		 * 'clen' in duk_hstring.
+		 *
+		 * The flag is set lazily for RAM strings.
 		 */
 		DUK_ASSERT(!DUK_HSTRING_HAS_ASCII(res));
-		if (DUK_LIKELY(clen == blen)) {
-			/* ASCII strings can't be symbol strings. */
-			DUK_HSTRING_SET_ASCII(res);
-		}
 	}
-#if defined(DUK_USE_HSTRING_CLEN)
-	DUK_HSTRING_SET_CHARLEN(res, clen);
-#endif
 
-	DUK_DDD(DUK_DDDPRINT("interned string, hash=0x%08lx, blen=%ld, clen=%ld, has_arridx=%ld, has_extdata=%ld",
+	DUK_DDD(DUK_DDDPRINT("interned string, hash=0x%08lx, blen=%ld, has_arridx=%ld, has_extdata=%ld",
 	                     (unsigned long) DUK_HSTRING_GET_HASH(res),
 	                     (long) DUK_HSTRING_GET_BYTELEN(res),
-	                     (long) DUK_HSTRING_GET_CHARLEN(res),
 	                     (long) (DUK_HSTRING_HAS_ARRIDX(res) ? 1 : 0),
 	                     (long) (DUK_HSTRING_HAS_EXTDATA(res) ? 1 : 0)));