For most strings the character length (`.length`) is not accessed. It's currently precomputed which slows down every string intern operation. Change that behavior so that the character length is computed lazily when it is first needed.

Tasks:
- [x] Change charlen and intern handling
- [x] Test effect in practice
- [x] Because of footprint impact, maybe make lazy/strict behavior optional => follow-up
- [x] Releases entry