@svaarala 2017-01-18
@fatcerberus This might also be interesting to test with the Typescript compiler. While this doesn't solve the fundamental problem in string accumulation loops (the repeated interning), it removes the charlen computation from the intermediate results which can make the accumulation run a bit faster.

For an extreme microbenchmark doing a char-by-char accumulation:

```
duktape@keppihevonen:~/duktape$ time ./duk.O2.master tests/perf/test-string-plain-concat.js 
real	0m0.910s
user	0m0.894s
sys	0m0.016s

duktape@keppihevonen:~/duktape$ time ./duk.O2 tests/perf/test-string-plain-concat.js 
real	0m0.538s
user	0m0.525s
sys	0m0.012s
```

Assuming 30% of time is spent in string concatenation and that about 40% of that would be optimized away, the net effect could be around 10-12%. Lots of variables in play, of course.