@fatcerberus 2017-01-18
Sure, I'll test it out - right after I re-add TypeScript support to Cell because I removed it when I switched to Babel (I mainly only wanted TS for the ES6 syntax). :)