@svaarala 2017-01-18
Oh sorry, I thought Typescript was in your normal workflow.

No need to go through that extra work -- I was just curious :) Maybe I should add a Babel and Typescript benchmark into the Duktape repo directly.