@fatcerberus 2017-01-18
I don't mind, so no worries.  I'm curious myself actually, because the profiling results for Babel differ so drastically from those for TypeScript that they clearly work very differently under the hood.