@fatcerberus 2017-01-18
Huh, TS (even without this branch) is actually much *faster* at ES6->ES5 transformation than Babel.  Good to know!