@svaarala 2017-01-18
Babel doesn't seem to be sensitive to this optimization: it's probably not spending a lot of time in string interning so the effect isn't visible. Test where babel-standalone converts an input file 10 times:
- 1.1.3: 13.4 seconds
- 1.2.5: 12.6 seconds
- 1.3.1: 11.4 seconds
- 1.4.0: 11.4 seconds
- 1.5.0: 11.2 seconds
- 2.0.0: 9.0 seconds
- master: 8.7 seconds
- master + perf options + PGO: 7.1 seconds
- this pull: 8.5 seconds
- this pull + perf options + PGO: 7.0 seconds

Master and this pull are most likely within the random variance so there may not be a measurable difference there.