@svaarala 2017-01-18
When optimizing small mechanisms it's always good to keep in mind that if the mechanism accounts for say 1% of the program execution, even making the mechanism 10x faster may not be measurable directly because the effect is a fraction of that 1%. Even so, doing ten small optimizations eventually turns up in application benchmarks too, so chipping away at the individual parts still matters.

This is definitely the case here - the effect is visible for string intern heavy code. Babel doesn't seem to be string heavy though.