@fatcerberus 2017-01-18
It seems that TypeScript isn't string-heavy anymore either.  My original profiling run was done with I think TypeScript 1.8, so maybe they've overhauled the compiler since then.  The profiling results now look much closer to Babel's.  Bytecode dispatch dominates (`duk__js_execute_bytecode_inner()` accounts for 19.46% of exclusive samples), another ~17% is spent in system calls (ntdll), and the rest are under 5%.

Interestingly, the profile summary tells me that the dominant hot path is:
```
duk_handle_call_inner
    duk_js_execute_bytecode
        duk__js_execute_bytecode_inner (19% exclusive)
            (HOT) duk_js_push_closure (30% inclusive)
            (HOT) duk_hobject_getprop (10% inclusive)
```

So it seems that this branch won't help TypeScript compilation at all, if the profiling run is anything to go by.