@svaarala 2017-01-18
That profile looks pretty normal in a sense. The closure push percentage is interesting -- it's most likely caused by there being a lot of functions with inner functions (callbacks and such). They are not handled very efficiently at present so maybe that percentage could be lower. For example, the property table could be copied as a byte copy rather than one property at a time as is currently done.

Having a microbenchmark for inner function creation would be interesting to track, I'll add one right away.