@fatcerberus 2017-01-18
The 17% for syscalls is interesting too (to me at least).  I assume those are memory allocations, since I can't imagine why else Duktape would need to call into the system.