@svaarala 2017-01-18
Memory allocations, or maybe setjmp/longjmp, string formatting, etc? Depends on what counts as a platform call.