@fatcerberus 2017-01-18
By system calls I specifically meant the "[ntdll.dll]" entries in the call list.  In other words calling into the kernel.

By the way, I don't know if this is interesting at all, but at the bottom of my profile summary:

```
Functions doing most individual work (exclusive samples):
duk__js_execute_bytecode_inner (20.68%)
[ntdll.dll] (16.12%)
duk_set_top (5.22%)
duk_hobject_find_existing_entry (4.21%)
duk_hobject_getprop (3.48%)
```