@svaarala 2017-01-18
I don't work much with Windows - but do malloc() and friends call into the kernel? I would assume they'd call into a userspace runtime which would only call into the kernel occasionally to e.g. map more pages.