@svaarala 2017-01-18
Running `strings ntdll.dll` shows that it might provide at least setjmp/longjmp.

That'd be easy to prove/disprove in practice - if you enable `DUK_USE_CPP_EXCEPTIONS` they would no longer be called :-)