@fatcerberus 2017-01-18
`malloc()` seems to call `HeapAlloc()` directly, which is part of the Win32 API.  Whether that involves an actual kernel call every time is an open question, the Win32 API is kind of a black box :)