@fatcerberus 2017-01-18
Anyway, yeah: TypeScript transpilation is unaffected by this pull.