@fatcerberus 2017-01-18
I think I might have stumbled upon a really pathological case with TypeScript.  For one particular source file in my game, TypeScript takes a long time (7-8 seconds) to compile it, while almost everything else is done in < 1s.

That in itself is interesting enough, but what's REALLY interesting is that the profile looks a lot different depending on whether refcounting is enabled or not.  With refcounting OFF, `duk_hobject_define_property_internal()` dominates with 28.50% of inclusive samples, and `duk_hobject_compact_props()` has 15.63% of inclusive samples.

With refcounting ON, `duk_js_push_closure()` dominates (34.31% inclusive), with `duk_hobject_getprop()` coming in second (11.50% inclusive).

In both cases the file takes a long time for TypeScript to process.