@fatcerberus 2017-01-18
On a sidenote: This is a case where I really wish Duktape had some sort of profiling framework in place so I could profile at the Ecmascript source level. :)